update :edit + delete JPA

+++ src/main/java/test/service/ProductService.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import test.product.Product;
 import test.repository.ProductRepository;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -14,17 +13,26 @@ public class ProductService {
     private ProductRepository repo;
 
     public List<Product> listAll(){
+
         return (List<Product>) repo.findAll();
     }
     public void save(Product product){
+
         repo.save(product);
     }
-    public Product get(Long id) {
-        return repo.findById(id).get();
+    public Product get(Long id) throws ProductNotFoundException {
+       Optional<Product> result= repo.findById(id);
+       if(result.isPresent()){
+           return result.get();
+       }
+       throw new ProductNotFoundException("Khong tim thay product voi ID "+id);
     }
 
-    public void delete(Long id){
+    public void delete(Long id) throws ProductNotFoundException {
+        Long count = repo.countById(id);
+        if(count == null || count == 0){
+            throw new ProductNotFoundException("Khong tim thay product voi ID "+id);
+        }
         repo.deleteById(id);
     }
-
 }
